linux: fix fallback check
I got it wrong when writing the doc: in case of fallback, the
getsockopt(MPTCP_INFO) syscall will always return an error,
MPTCP_INFO_FLAG_FALLBACK will never be set -- it is only set when the
info are retrieved via Netlink.

So we can simplify this, and partly revert PR #1.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>